+++ CHANGELOG.md
@@ -72,6 +72,7 @@
 * [BUGFIX] QueryFrontend: fixed query_range requests when query has `start` equals to `end`. #4877
 * [BUGFIX] AlertManager: fixed issue introduced by #4495 where templates files were being deleted when using alertmanager local store. #4890
 * [BUGFIX] Ingester: fixed incorrect logging at the start of ingester block shipping logic. #4934
+* [BUGFIX] Storage/Bucket: fixed global mark missing on deletion. #4949
 
 ## 1.13.0 2022-07-14